feat(useParseTemplateData): string 데이터를 Card 데이터로 파싱하는 로직 hook으로 분리
